# frontend/src/api/openapi.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "FailResponseFromError_InvalidTokenError_": {
        "$ref": "#/components/schemas/FailResponse_INVALID_TOKEN.undefined_",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
...
      "AuthFailResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/FailResponseFromError_NoTokenProvidedError_"
          },
          {
            "$ref": "#/components/schemas/FailResponseFromError_InvalidTokenError_"
          },
          {
            "$ref": "#/components/schemas/FailResponseFromError_TokenUserNotFoundError_"
          },
          {
            "$ref": "#/components/schemas/FailResponseFromError_AuthenticationVersionMismatchError_"
          }
        ],
        "description": "Tipo de respuesta para errores de autenticación JWT (401).\r\nCubre todos los errores lanzados por expressAuthentication:\r\n- NO_TOKEN_PROVIDED: No se proporcionó token (con reason en details)\r\n- INVALID_TOKEN: Token inválido o expirado\r\n- TOKEN_USER_NOT_FOUND: Usuario del token ya no existe (con userId en details)\r\n- AUTH_VERSION_MISMATCH: Sesión invalidada (con userId, currentVersion, tokenVersion en details)"
      },
...
      "SuccessResponse_LoginResponseData_": {
        "description": "Wrapper de éxito JSend (aplicado por el middleware global).",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "nullable": false
          },
          "data": {
            "$ref": "#/components/schemas/LoginResponseData"
          }
        },
        "required": [
          "status",
          "data"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "LoginValidationFailResponse": {
        "$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_BodyPath_LoginRequest___"
      },
...
      "FailResponseFromError_InvalidCredentialsError_": {
        "$ref": "#/components/schemas/FailResponse_INVALID_CREDENTIALS._identifier-string__",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
      "LoginRequest": {
        "properties": {
          "identifier": {
            "type": "string",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "responseType": {
            "type": "string",
            "enum": [
              "cookie",
              "json"
            ],
            "description": "Indicates how the token should be returned.",
            "default": "json"
          }
        },
        "required": [
          "identifier",
          "password"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "GoogleLoginValidationFailResponse": {
        "$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_BodyPath_GoogleLoginRequest___"
      },
      "GoogleLoginRequest": {
        "properties": {
          "credential": {
            "type": "string"
          },
          "responseType": {
            "type": "string",
            "enum": [
              "cookie",
              "json"
            ],
            "description": "Indicates how the token should be returned.",
            "default": "json"
          }
        },
        "required": [
          "credential"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "SuccessResponse_MessageResponseData_": {
        "description": "Wrapper de éxito JSend (aplicado por el middleware global).",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "nullable": false
          },
          "data": {
            "$ref": "#/components/schemas/MessageResponseData"
          }
        },
        "required": [
          "status",
          "data"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "GoogleConnectValidationFailResponse": {
        "$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_BodyPath_GoogleConnectRequest___"
      },
...
      "FailResponseFromError_GoogleAccountAlreadyLinkedError_": {
        "$ref": "#/components/schemas/FailResponse_GOOGLE_ACCOUNT_ALREADY_LINKED._email-string__",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
      "GoogleConnectRequest": {
        "properties": {
          "credential": {
            "type": "string"
          }
        },
        "required": [
          "credential"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "FailResponseFromError_CannotDisconnectGoogleWithoutPasswordError_": {
        "$ref": "#/components/schemas/FailResponse_CANNOT_DISCONNECT_GOOGLE_WITHOUT_PASSWORD.undefined_",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
...
      "ChangePasswordRequestValidationFailResponse": {
        "$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_BodyPath_ChangePasswordRequest___"
      },
      "ChangePasswordValidationFailResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ChangePasswordRequestValidationFailResponse"
          },
          {
            "$ref": "#/components/schemas/DatabaseValidationFailResponse"
          }
        ]
      },
...
      "ChangePasswordErrorResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/AuthFailResponse"
          },
          {
            "$ref": "#/components/schemas/FailResponseFromError_InvalidPasswordError_"
          }
        ]
      },
...
      "FailResponseFromError_NewPasswordSameAsOldError_": {
        "$ref": "#/components/schemas/FailResponse_NEW_PASSWORD_SAME_AS_OLD.undefined_",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
...
      "FailResponseFromError_LoginUserNotFoundError_": {
        "$ref": "#/components/schemas/FailResponse_USER_NOT_FOUND._identifier-string__",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
      "ChangePasswordRequest": {
        "properties": {
          "oldPassword": {
            "type": "string",
            "minLength": 8
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "SetPasswordValidationFailResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/SetPasswordRequestValidationFailResponse"
          },
          {
            "$ref": "#/components/schemas/DatabaseValidationFailResponse"
          }
        ]
      },
...
      "FailResponseFromError_PasswordAlreadySetError_": {
        "$ref": "#/components/schemas/FailResponse_PASSWORD_ALREADY_SET.undefined_",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
      "SetPasswordRequest": {
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "password"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "ForgotPasswordValidationFailResponse": {
        "$ref": "#/components/schemas/ForgotPasswordRequestValidationFailResponse"
      },
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9]))\\.){3}(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])$"
          }
        },
        "required": [
          "email"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "ResetPasswordValidationFailResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ResetPasswordRequestValidationFailResponse"
          },
          {
            "$ref": "#/components/schemas/DatabaseValidationFailResponse"
          }
        ]
      },
...
      "FailResponseFromError_ResetTokenInvalidOrExpiredError_": {
        "$ref": "#/components/schemas/FailResponse_RESET_TOKEN_INVALID_OR_EXPIRED.undefined_",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "jwt-optional": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "@flaights/backend",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "contact": {}
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_LoginResponseData_"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_InvalidCredentialsError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Inicia sesión con un identificador (email o username) y contraseña.",
        "tags": [
          "Auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/auth/login/google": {
      "post": {
        "operationId": "LoginWithGoogle",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_LoginResponseData_"
                }
              }
            }
          },
          "401": {
            "description": "Token de Google inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_InvalidTokenError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleLoginValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Inicia sesión o registra usuario mediante Google OAuth.",
        "tags": [
          "Auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginRequest"
              }
            }
          }
        }
      }
    },
    "/auth/google/connect": {
      "post": {
        "operationId": "ConnectGoogle",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "409": {
            "description": "Cuenta de Google ya vinculada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_GoogleAccountAlreadyLinkedError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleConnectValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Vincula la cuenta de Google al usuario actual.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleConnectRequest"
              }
            }
          }
        }
      }
    },
    "/auth/google/disconnect": {
      "post": {
        "operationId": "DisconnectGoogle",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
                }
              }
            }
          },
          "400": {
            "description": "Debe establecer una contraseña primero",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_CannotDisconnectGoogleWithoutPasswordError_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          }
        },
        "description": "Desvincula la cuenta de Google del usuario actual.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": []
      }
    },
...
    "/auth/logoutAll": {
      "post": {
        "operationId": "LogoutAll",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          }
        },
        "description": "Cierra todas las sesiones activas del usuario (invalida todos los tokens).",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": []
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "ChangePassword",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
                }
              }
            }
          },
          "400": {
            "description": "Nueva contraseña igual a la anterior",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_NewPasswordSameAsOldError_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado o contraseña incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_LoginUserNotFoundError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Cambia la contraseña del usuario autenticado.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        }
      }
    },
    "/auth/set-password": {
      "post": {
        "operationId": "SetPassword",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
                }
              }
            }
          },
          "400": {
            "description": "Contraseña ya establecida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_PasswordAlreadySetError_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetPasswordValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Establece la contraseña por primera vez para el usuario autenticado (si no tiene una).",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordRequest"
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "ForgotPassword",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Solicita un email de recuperación de contraseña.\r\nPor seguridad, siempre devuelve éxito aunque el email no exista.",
        "tags": [
          "Auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "ResetPassword",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_MessageResponseData_"
                }
              }
            }
          },
          "400": {
            "description": "Token inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_ResetTokenInvalidOrExpiredError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Restablece la contraseña usando un token de recuperación.",
        "tags": [
          "Auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}